seperated functions into files issue #98
